+++ ru.taximaxim.codekeeper.core/src/main/java/ru/taximaxim/codekeeper/core/PgDiff.java
@@ -410,6 +410,7 @@ public class PgDiff {
             }
         }
         depRes.recreateDrops();
+        depRes.removeExtraActions();
     }